+++ tests/tools.py
@@ -51,4 +51,4 @@ def include_views(config):
 @view_config(route_name="secret", permission="super_high", renderer='json')
 def secret_view(request):
     """Dummy view with redirect to login."""
-    return dict()
+    return {}